[NFC] Make CHIRRTL more normal in preparation for moving some stuff in
